TestFramework: Move classes to IssueValidation namespace (#8591)
