+++ analyzers/tests/SonarAnalyzer.TestFramework.Test/Verification/IssueValidation/IssueLocationCollectorTest.GetExpectedIssueLocations.cs
@@ -19,6 +19,7 @@
  */
 
 using Microsoft.CodeAnalysis.Text;
+using SonarAnalyzer.TestFramework.Verification.IssueValidation;
 
 namespace SonarAnalyzer.Test.TestFramework.Tests
 {
