+++ analyzers/tests/SonarAnalyzer.TestFramework.Test/Verification/IssueValidation/IssueLocationCollectorTest.GetIssueLocations.cs
@@ -18,6 +18,8 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
+using SonarAnalyzer.TestFramework.Verification.IssueValidation;
+
 namespace SonarAnalyzer.Test.TestFramework.Tests
 {
     public partial class IssueLocationCollectorTest
