+++ analyzers/tests/SonarAnalyzer.TestFramework.Test/Verification/IssueValidation/IssueLocationCollectorTest.MergeLocations.cs
@@ -18,7 +18,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using IssueLocation = SonarAnalyzer.Test.TestFramework.IssueLocationCollector.IssueLocation;
+using SonarAnalyzer.TestFramework.Verification.IssueValidation;
 
 namespace SonarAnalyzer.Test.TestFramework.Tests
 {
