+++ analyzers/tests/SonarAnalyzer.TestFramework/Verification/CodeFixVerifier.cs
@@ -21,6 +21,7 @@
 using System.IO;
 using Microsoft.CodeAnalysis.CodeActions;
 using Microsoft.CodeAnalysis.CodeFixes;
+using SonarAnalyzer.TestFramework.Verification.IssueValidation;
 
 namespace SonarAnalyzer.Test.TestFramework;
